+++ .gitignore
@@ -10,6 +10,7 @@ test/tmp
 test/version_tmp
 tmp
 _Store
+.DS_Store
 *~
 *#
 .#*
