+++ test/cookbooks/test/recipes/app.rb
@@ -1,8 +1,8 @@
 #
 # Cookbook:: test
-# Recipe:: default
+# Recipe:: app
 #
-# Copyright:: 2017, The Authors
+# copyright: 2017, Chef Software, Inc.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
